Move Locale to Localization.py
