+++ EpikCord/locales.py
@@ -1,8 +1,38 @@
+from enum import Enum
 from typing import Dict
 
-from .utils import Locale
-
-
+class Locale(Enum):
+    INDONESIAN = "id"
+    DANISH = "da"
+    GERMAN = "de"
+    ENGLISH_UK = "en-GB"
+    ENGLISH_US = "en-US"
+    SPANISH = "es-ES"
+    FRENCH = "fr"
+    CROATIAN = "hr"
+    ITALIAN = "it"
+    LITHUANIAN = "lt"
+    HUNGARIAN = "hu"
+    DUTCH = "nl"
+    NORWEGIAN = "no"
+    POLISH = "pl"
+    PORTUGUESE_BRAZILIAN = "pt-BR"
+    ROMANIAN_ROMANIA = "ro"
+    FINNISH = "fi"
+    SWEDISH = "sv-SE"
+    VIETNAMESE = "vi"
+    TURKISH = "tr"
+    CZECH = "cs"
+    GREEK = "el"
+    BULGARIAN = "bg"
+    RUSSIAN = "ru"
+    UKRAINIAN = "uk"
+    HINDI = "hi"
+    THAI = "th"
+    CHINESE_CHINA = "zh-CN"
+    JAPANESE = "ja"
+    CHINESE_TAIWAN = "zh-TW"
+    KOREAN = "ko"
 class Localization:
     def __init__(self, locale: Locale, value: str):
         self.locale = locale
