+++ EpikCord/utils/__init__.py
@@ -6,7 +6,6 @@ from .enums import (
     HTTPCodes,
     InteractionType,
     JSONErrorCodes,
-    Locale,
     OpCode,
     PremiumType,
     StatusCode,
@@ -51,7 +50,6 @@ __all__ = (
     "ApplicationCommandType",
     "TeamMembershipState",
     "ApplicationCommandOptionType",
-    "Locale",
     "PremiumType",
     "add_file",
     "cancel_tasks",
